Create activity.helper and extend activity.model with it
